[add + fix] コンポーネントの使い方についてコメントを追加 LAページの差分を無くす
